- diary Team added as json files - password check with json file - login error message - login screen hidden when logged in
