--- Diary/Diary_Eintraege/Rene.json
@@ -1,1 +1,0 @@
-{"name":"Rene","rolle":"Websiteguy","messageList":[{"Datum":"05.02.2024","name":"rene","beschreibung":"website weiterentwickelt"}]}
