+++ Diary/Diary_Eintraege/admin.json
@@ -1,0 +1,1 @@
+{"name":"admin","rolle":"","password":"password","messageList":[]}
