+++ Diary/Diary_Eintraege/alex.json
@@ -1,0 +1,1 @@
+{"name":"alex","rolle":"","password":"password","messageList":[]}
