+++ Diary/Diary_Eintraege/patrick.json
@@ -1,0 +1,1 @@
+{"name":"patrick","rolle":"","password":"password","messageList":[]}
