+++ Diary/Diary_Eintraege/pillip.json
@@ -1,0 +1,1 @@
+{"name":"pillip","rolle":"","password":"password","messageList":[]}
