+++ Diary/Diary_Eintraege/rene.json
@@ -1,0 +1,1 @@
+{"name":"rene","rolle":"Website-Guy","password":"!?","messageList":[]}
