+++ Diary/functions_diary.js
@@ -1,3 +1,7 @@
+//import("../funktionen.js")
+//document.getElementById("diary-display").classList.add("fade-in");
+
+
 document.addEventListener("DOMContentLoaded", function() {
     const form = document.getElementById("loginForm");
     form.onsubmit = function(e) {
@@ -8,7 +12,23 @@ document.addEventListener("DOMContentLoaded", function() {
             body: formData
         })
             .then(response => response.text())
-            .then(data => document.getElementById("loginResult").innerText = data)
+            .then(data => processData(data))
             .catch(error => console.error('Error:', error));
     };
 });
+
+function  processData(data){
+    if(data === "true"){
+        document.getElementById("loginForm").style.display = "none";
+        document.getElementById("diary-display").style.display = "block";
+        //toggleFadeIn(document.getElementById("diary-display"));
+        loadPerson();
+    }else{
+        document.getElementById("loginResult").innerHTML = "wrong username or password";
+    }
+
+}
+
+function loadPerson(username, password){
+
+}
