+++ Diary/style_diary.css
@@ -1,4 +1,4 @@
-#loginForm{ /*good to copy for home!!!*/
+.diary-container{ /*good to copy for home!!!*/
     display: flex;
     flex-direction: row;
     align-items: center;
@@ -44,3 +44,24 @@
     box-shadow: 0 4px 8px rgba(0,0,0,0.1);
     border-radius: 8px;
     background-color: #eae8df;}
+
+
+#loginResult{
+    color: red;
+}
+/* Display */
+#diary-display{
+    display: none;
+}
+
+.fade-in {
+    opacity: 0;
+    visibility: hidden;
+    transition: opacity 0.5s, visibility 0s linear 0.5s; /* Verzögert die Änderung der Sichtbarkeit bis zum Ende des Übergangs */
+}
+
+.fade-in.visible {
+    opacity: 1;
+    visibility: visible;
+    transition: opacity 1s; /* Macht den Übergang sofort sichtbar */
+}
